+++ app/imports/templates/modals/backup.html
@@ -1,0 +1,9 @@
+<template name="modals_backup">
+  <h1> Backup your bids</h1>
+
+  <p>This allows you to give them to someone else to reveal them, or restore them in case of data loss <br></p>
+
+  <div class="dapp-modal-buttons no-margins">
+    <button class="ok dapp-primary-button download">Download bids</button>
+  </div>
+</template>
