+++ app/imports/templates/modals/restore.html
@@ -1,20 +1,4 @@
-<template name="modals_backupRestore">
-  <div>
-    {{> Template.dynamic template=view }}
-  </div>
-</template>
-
-<template name="backup">
-  <h1> Backup your bids</h1>
-
-  <p>This allows you to give them to someone else to reveal them, or restore them in case of data loss <br></p>
-
-  <div class="dapp-modal-buttons no-margins">
-    <button class="ok dapp-primary-button download">Download bids</button>
-  </div>
-</template>
-
-<template name="restore">
+<template name="modals_restore">
   <h1>Import Bids</h1>
   <p>
     This allows you to reveal bids made in other computers. Your current bids won't get removed. <br>
